Use the resource when receiving a message from an anonymous account
I realized that the nickname option for anonymous accounts only applies
to MUCs.  The nickname is sent with every message as per XEP-0172
suggestion but Jappix Desktop nor Pidgin set the nickname using this
method, making the UUID the visible name.

The only part of the full JID that's modifiable for anonymous accounts
is the resource, so this patch checks if the JID is anonymous (the
username is an UUID) and if so, uses the resource.