mam: Assume MAM is always available.
Post v0.1 MAM is a bare-JID cap. Existing code assumes that this is a server JID
cap.

Because the conceptual change in who advertises support for the archive is
substantial, more work would be required to query caps for a bare JID, for
remote user's bare JID, MUC JID, etc.